requirements.txt
langchain
pandas 
langchain_experimental
google-generativeai 
langchain-google-genai
python-dotenv
tabulate
Streamlit
numpy